- getDatasByOrderNumber checks now for payment_currency > "0" to get the correct entry - paypal does not try a refund for order status cancelled - added config ordersAddOnly which prevents deletion of order item.
git-svn-id: https://dev.virtuemart.net/svn/virtuemart@9761 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ branches/com_virtuemart.3.0.12.4/administrator/components/com_virtuemart/models/orders.php
@@ -2179,21 +2179,14 @@ vmdebug('my prices',$data);
 		//answer : when creating an order (P) the stock is reserved
 		$this->handleStockAfterStatusChangedPerProduct('X', $item->order_status, $item,$item->product_quantity);
 
-
-
-		//store history
+		//take data for history
 		$data = $item->getProperties();
-		$data['action'] = 'deleted';
-		$tableHist = $this->getTable('order_item_histories');
 
-		$tableHist->bindChecknStore($data);
+		if (!VmConfig::get('ordersAddOnly',false) and $item->delete($orderLineId)) {
+			$data['action'] = 'deleted';
+			$tableHist = $this->getTable('order_item_histories');
 
-
-		if ($item->delete($orderLineId)) {
-			/*$q = "DELETE FROM `#__virtuemart_order_histories`
-			WHERE `virtuemart_order_id`=".$id;
-			$this->_db->setQuery($q);
-			$this->_db->execute();*/
+			$tableHist->bindChecknStore($data);
 
 			$q = "DELETE FROM `#__virtuemart_order_calc_rules`
 			WHERE `virtuemart_order_item_id`=".$orderLineId;
